+++ src/game/Object/Player.cpp
@@ -3939,125 +3939,6 @@ void Player::SetHover(bool enable)
 }
 
 /**
- * @brief Registers a joined chat channel on the player.
- *
- * @param c The channel that was joined.
- */
-void Player::JoinedChannel(Channel* c)
-{
-    m_channels.push_back(c);
-}
-
-/**
- * @brief Removes a chat channel from the player's joined channel list.
- *
- * @param c The channel that was left.
- */
-void Player::LeftChannel(Channel* c)
-{
-    m_channels.remove(c);
-}
-
-/**
- * @brief Leaves and cleans up all channels currently joined by the player.
- */
-void Player::CleanupChannels()
-{
-    while (!m_channels.empty())
-    {
-        Channel* ch = *m_channels.begin();
-        m_channels.erase(m_channels.begin());               // remove from player's channel list
-        ch->Leave(this, false);                             // not send to client, not remove from player's channel list
-        if (ChannelMgr* cMgr = channelMgr(GetTeam()))
-        {
-            cMgr->LeftChannel(ch->GetName()); // deleted channel if empty
-        }
-    }
-    DEBUG_LOG("Player: channels cleaned up!");
-}
-
-/**
- * @brief Updates built-in local channels after a zone change.
- *
- * @param newZone The new area identifier used for localized channel names.
- */
-void Player::UpdateLocalChannels(uint32 newZone)
-{
-    if (m_channels.empty())
-    {
-        return;
-    }
-
-    AreaTableEntry const* current_zone = GetAreaEntryByAreaID(newZone);
-    if (!current_zone)
-    {
-        return;
-    }
-
-    ChannelMgr* cMgr = channelMgr(GetTeam());
-    if (!cMgr)
-    {
-        return;
-    }
-
-    std::string current_zone_name = current_zone->area_name[GetSession()->GetSessionDbcLocale()];
-
-    for (JoinedChannelsList::iterator i = m_channels.begin(), next; i != m_channels.end(); i = next)
-    {
-        next = i; ++next;
-
-        // skip non built-in channels
-        if (!(*i)->IsConstant())
-        {
-            continue;
-        }
-
-        ChatChannelsEntry const* ch = GetChannelEntryFor((*i)->GetChannelId());
-        if (!ch)
-        {
-            continue;
-        }
-
-        if ((ch->flags & 4) == 4)                           // global channel without zone name in pattern
-        {
-            continue;
-        }
-
-        //  new channel
-        char new_channel_name_buf[100];
-        snprintf(new_channel_name_buf, 100, ch->pattern[m_session->GetSessionDbcLocale()], current_zone_name.c_str());
-        Channel* new_channel = cMgr->GetJoinChannel(new_channel_name_buf, ch->ChannelID);
-
-        if ((*i) != new_channel)
-        {
-            new_channel->Join(this, "");                    // will output Changed Channel: N. Name
-
-            // leave old channel
-            (*i)->Leave(this, false);                       // not send leave channel, it already replaced at client
-            std::string name = (*i)->GetName();             // store name, (*i)erase in LeftChannel
-            LeftChannel(*i);                                // remove from player's channel list
-            cMgr->LeftChannel(name);                        // delete if empty
-        }
-    }
-    DEBUG_LOG("Player: channels cleaned up!");
-}
-
-/**
- * @brief Leaves the currently joined looking-for-group channel, if any.
- */
-void Player::LeaveLFGChannel()
-{
-    for (JoinedChannelsList::iterator i = m_channels.begin(); i != m_channels.end(); ++i)
-    {
-        if ((*i)->IsLFG())
-        {
-            (*i)->Leave(this);
-            break;
-        }
-    }
-}
-
-/**
  * @brief Attempts to improve defense skill and refresh defense-derived bonuses.
  */
 void Player::UpdateDefense()
@@ -4255,183 +4136,6 @@ void Player::UpdateCombatSkills(Unit* pVictim, WeaponAttackType attType, bool de
     {
         return;
     }
-}
-
-/* Called from Player::SendInitialPacketsBeforeAddToMap */
-/**
- * @brief Sends the player's initial action bar state to the client.
- */
-void Player::SendInitialActionButtons() const
-{
-    DETAIL_LOG("Initializing Action Buttons for '%u'", GetGUIDLow());
-
-    /* Initiate packet with size 4 bytes per action button */
-    WorldPacket data(SMSG_ACTION_BUTTONS, (MAX_ACTION_BUTTONS * 4));
-
-    /* For each possible action button the player could have */
-    for (uint8 button = 0; button < MAX_ACTION_BUTTONS; ++button)
-    {
-        /* Try and get each action button the player could have */
-        ActionButtonList::const_iterator itr = m_actionButtons.find(button);
-
-        /* If the button is valid and not deleted */
-        if (itr != m_actionButtons.end() && itr->second.uState != ACTIONBUTTON_DELETED)
-        {
-            /* Send the data */
-            data << uint32(itr->second.packedData);
-        }
-        else
-        {
-            /* Nothing to send, so just send 0 */
-            data << uint32(0);
-        }
-    }
-
-    GetSession()->SendPacket(&data);
-    DETAIL_LOG("Action Buttons for '%u' Initialized", GetGUIDLow());
-}
-
-/**
- * @brief Validates action bar data before it is stored or loaded.
- *
- * @param button The action bar button index.
- * @param action The action identifier assigned to the button.
- * @param type The action button type.
- * @param player The player being validated for, or null for template data.
- * @return True if the action button data is valid; otherwise, false.
- */
-bool Player::IsActionButtonDataValid(uint8 button, uint32 action, uint8 type, Player* player)
-{
-    if (button >= MAX_ACTION_BUTTONS)
-    {
-        if (player)
-        {
-            sLog.outError("Action %u not added into button %u for player %s: button must be < %u", action, button, player->GetName(), MAX_ACTION_BUTTONS);
-        }
-        else
-        {
-            sLog.outError("Table `playercreateinfo_action` have action %u into button %u : button must be < %u", action, button, MAX_ACTION_BUTTONS);
-        }
-        return false;
-    }
-
-    if (action >= MAX_ACTION_BUTTON_ACTION_VALUE)
-    {
-        if (player)
-        {
-            sLog.outError("Action %u not added into button %u for player %s: action must be < %u", action, button, player->GetName(), MAX_ACTION_BUTTON_ACTION_VALUE);
-        }
-        else
-        {
-            sLog.outError("Table `playercreateinfo_action` have action %u into button %u : action must be < %u", action, button, MAX_ACTION_BUTTON_ACTION_VALUE);
-        }
-        return false;
-    }
-
-    switch (type)
-    {
-        case ACTION_BUTTON_SPELL:
-        {
-            SpellEntry const* spellProto = sSpellStore.LookupEntry(action);
-            if (!spellProto)
-            {
-                if (player)
-                {
-                    sLog.outError("Spell action %u not added into button %u for player %s: spell not exist", action, button, player->GetName());
-                }
-                else
-                {
-                    sLog.outError("Table `playercreateinfo_action` have spell action %u into button %u: spell not exist", action, button);
-                }
-                return false;
-            }
-
-            if (player)
-            {
-                if (!player->HasSpell(spellProto->Id))
-                {
-                    sLog.outError("Spell action %u not added into button %u for player %s: player don't known this spell", action, button, player->GetName());
-                    return false;
-                }
-                else if (IsPassiveSpell(spellProto))
-                {
-                    sLog.outError("Spell action %u not added into button %u for player %s: spell is passive", action, button, player->GetName());
-                    return false;
-                }
-            }
-            break;
-        }
-        case ACTION_BUTTON_ITEM:
-        {
-            if (!ObjectMgr::GetItemPrototype(action))
-            {
-                if (player)
-                {
-                    sLog.outError("Item action %u not added into button %u for player %s: item not exist", action, button, player->GetName());
-                }
-                else
-                {
-                    sLog.outError("Table `playercreateinfo_action` have item action %u into button %u: item not exist", action, button);
-                }
-                return false;
-            }
-            break;
-        }
-        default:
-            break;                                          // other cases not checked at this moment
-    }
-
-    return true;
-}
-
-/**
- * @brief Adds or updates an action bar button entry.
- *
- * @param button The action bar button index.
- * @param action The action identifier to bind.
- * @param type The action button type.
- * @return The updated action button, or null if validation fails.
- */
-ActionButton* Player::addActionButton(uint8 button, uint32 action, uint8 type)
-{
-    if (!IsActionButtonDataValid(button, action, type, this))
-    {
-        return NULL;
-    }
-
-    // it create new button (NEW state) if need or return existing
-    ActionButton& ab = m_actionButtons[button];
-
-    // set data and update to CHANGED if not NEW
-    ab.SetActionAndType(action, ActionButtonType(type));
-
-    DETAIL_LOG("Player '%u' Added Action '%u' (type %u) to Button '%u'", GetGUIDLow(), action, uint32(type), button);
-    return &ab;
-}
-
-/**
- * @brief Removes an action bar button entry.
- *
- * @param button The action bar button index to remove.
- */
-void Player::removeActionButton(uint8 button)
-{
-    ActionButtonList::iterator buttonItr = m_actionButtons.find(button);
-    if (buttonItr == m_actionButtons.end())
-    {
-        return;
-    }
-
-    if (buttonItr->second.uState == ACTIONBUTTON_NEW)
-    {
-        m_actionButtons.erase(buttonItr); // new and not saved
-    }
-    else
-    {
-        buttonItr->second.uState = ACTIONBUTTON_DELETED; // saved, will deleted at next save
-    }
-
-    DETAIL_LOG("Action Button '%u' Removed from Player '%u'", button, GetGUIDLow());
 }
 
 /**
